project env configs done

+++ src/tasks/tasks.service.ts
@@ -1,6 +1,7 @@
 /* eslint-disable prettier/prettier */
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTaskFilterDto } from './dto/get-task-filter.dto';
 import { TaskStatus } from './task-status.enum';
@@ -9,40 +10,44 @@ import { TaskRepository } from './task.repository';
 
 @Injectable()
 export class TasksService {
-
-  constructor (
+  constructor(
     @InjectRepository(TaskRepository)
-    private taskRepository: TaskRepository
-   ) {}
+    private taskRepository: TaskRepository,
+  ) {}
 
-  async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-    return await this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+    return await this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const task = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const task = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
 
-    if(!task){
-      throw new NotFoundException(`Task with id:${id} does not exist in the system`,);
+    if (!task) {
+      throw new NotFoundException(
+        `Task with id:${id} does not exist in the system`,
+      );
     }
-      
 
     return task;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
-    if(result.affected === 0){
-      throw new NotFoundException(`Task with id:${id} does not exist in the system`,);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
+    if (result.affected === 0) {
+      throw new NotFoundException(
+        `Task with id:${id} does not exist in the system`,
+      );
     }
   }
 
-  async updateTask(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTask(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
